almost fixed unit test of useClickOutside
